Replace all classes and functions in single files
