+++ person.cpp
@@ -1,0 +1,5 @@
+#include "person.h"
+
+Person::Person(string inName, int inAge): name(inName), age(inAge)
+{
+}
